Finished

+++ pages/api/products/[id].js
@@ -5,7 +5,9 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+  const token = cookies.token;
   await dbConnect();
 
   if (method === 'GET') {
@@ -17,6 +19,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === 'PUT') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json({ message: 'Not authenticated' });
+    }
     try {
       const product = await Product.create(req.body);
       res
@@ -27,6 +32,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === 'DELETE') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json({ message: 'Not authenticated' });
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json({ message: 'Product deleted successfully!' });
